Create config file for gitsigns
